combine front matter and handlebar templates
The basic idea is, that neither html nor hbs file are bothered by a front matter configuration and you could let them be processed through matter() with and without frontmatter.
It also ignores if a frontmatter structure is somewhere else in the template.

So this would be an option for templating as well